refactor: factor out WebViewEmitter
