+++ Sources/PagecallSDK/MiController.swift
@@ -5,7 +5,6 @@
 //  Created by Jaeseong Seo on 2023/03/16.
 //
 
-import Foundation
 import Mediasoup
 import WebRTC
